Add file checking to uninstall script.

+++ uninstall.sh
@@ -6,6 +6,12 @@ unlink ~/.vimrc
 unlink ~/.bashrc
 unlink ~/.tmux.conf
 
-mv ~/.vimrc.bak ~/.vimrc
-mv ~/.bashrc.bak ~/.bashrc
-mv ~/.tmux.conf.bak ~/.tmux.conf
+if [ -f ~/.vimrc.bak ]; then
+  mv ~/.vimrc.bak ~/.vimrc
+fi
+if [ -f ~/.bashrc.bak ]; then
+  mv ~/.bashrc.bak ~/.bashrc
+fi
+if [ -f ~/.tmux.conf.bak ]; then
+  mv ~/.tmux.conf.bak ~/.tmux.conf
+fi
